REFACTOR liblinear based estimators
BaseLibLinear -> LibLinearMixin
